requirements.txt
streamlit>=1.35.0
plotly>=6.0.0
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
requests>=2.31.0
